refactor: address teammate review feedback
- Move helpers (_write_env_file, INTERNAL_REGISTRY) and fixtures
  (agent_dir, agent_name, deployed_agent) to conftest.py so the test
  file contains only test functions (andrewdonheiser)
- Add broad `except Exception` fallback in deployed_agent fixture so
  unexpected errors surface clearly in CI (andrewdonheiser)
- Wrap EMBEDDING_DIMENSION int() cast with try/except and pytest.fail
  for a clear error message (tarun-etikala)
